early-boot-config: Add local defaults user data provider
Adds a standalone user data provider binary specifically for gathering
data from the local defaults file `/local/user_data_defaults.toml`.  The
binary is meant to be run by the main `early-boot-config` executable.
It logs to stderr, as it returns a `SettingsJson` object on stdout in
the case of success.  It returns an error and non-zero exit code in the
case of failure.