packages: Add 'early-boot-config-ec2-imds' package
Adds a new package for the `ec2-imds` user data provider.  The package
includes 2 subpackages, a '*-client' package which includes just the
binary, and a '*-provider' package which symlinks the binary to the .d
diredtory in which `early-boot-config` will look for providers.  The
default prefix in the '*-provider' package is '40-' do place it fourth
in order.