early-boot-config: Add logic to find and execute user provider binaries
Adds the logic to find and execute user data provider binaries in the
proper order.  Provider logs will be forwarded to assist in any
troubleshooting necessary.

The `SettingsJson` object forms the contract between `early-boot-config`
and the provider binaries.  Previously this object store user data as a
serialized string.  Since provider binaries serialize this object to
stdout, user data is now stored as a serde_json::Value to avoid
"double-serializing" the JSON.